Added in drawing functions for TetrisPieces.
Doesn't seem to work, but it could be a different issue.
Getting close to have a basic working Tetris game!

+++ include/game/block/tetris_block.h
@@ -68,7 +68,7 @@ extern BlockType tetris_block_get_type(TetrisBlock *self);
  * @param y The y-co-ordinate of the top left of the box. (Where 0,0 is top left of the screen)
  * @param length The width and height of the block.
  */
-extern void tetris_block_draw(TetrisBlock *self, int x, int y, int length);
+extern void tetris_block_draw(TetrisBlock *self, u16 x, u16 y, u16 length);
